refactor: Move logging to Vira.Lib
Since the logging part doesn't require application domain, so it
shouldn't be in Vira.App.

Incidentally, this makes Vira.Supervisor indepent of the rest of the app
(only dependent on Vira.Lib.*)
