+++ packages/vira/src/Vira/App.hs
@@ -9,6 +9,6 @@ where
 
 import Vira.App.AcidState as X
 import Vira.App.CLI as X
-import Vira.App.Logging as X (log)
 import Vira.App.Servant as X
 import Vira.App.Stack as X
+import Vira.Lib.Logging as X (log)
